requirements.txt
dearpygui~=1.8.0
numpy~=1.24.1
Pillow~=9.4.0